requirements.txt
groq==0.28.0
supabase==2.15.3
google-api-python-client==2.173.0
google-auth==2.40.3
flask==3.0.0
requests==2.31.0
pytest==8.3.4
python-dateutil==2.8.2
pytz==2024.1